requirements.txt
flask==3.1.0
groq==0.25.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
